// src/interfaces/flights.ts
export interface iLLBGflight {
	CheckInUrl: string;
	Counter: string;
	Airline: string;
	Flight: string;
	Terminal: string;
	Status: string;
	City: string;
	Country: string;
	StatusColor: string;
	ScheduledDateTime: string;
	ScheduledDate: string;
	ScheduledTime: string;
	UpdatedDateTime: string;
	UpdatedDate: string;
	UpdatedTime: string;
	CurrentCultureName: string;
}